+++ news/Dockerfile
@@ -1,0 +1,17 @@
+FROM node:alpine
+
+# set working directory
+WORKDIR /app
+
+# add `/app/node_modules/.bin` to $PATH
+ENV PATH /app/node_modules/.bin:$PATH
+
+# install and cache app dependencies
+COPY package.json /app/package.json
+COPY package-lock.json /app/package-lock.json
+
+RUN npm install
+COPY . ./
+
+# start app
+CMD ["npm", "start"]
