+++ news/src/App.js
@@ -8,13 +8,9 @@ function App() {
 
     return (
         <>
-            <div className = "collum1">
-                <div className = "header">
-                    <h1>Custom News</h1>
-                    <h2>News your way</h2>
-                </div>
-
-
+            <div className = "header">
+                <h1>Custom News</h1>
+                <h2>News your way</h2>
             </div>
             <div className = "news">
                 <h2>Search for keywords</h2>
