+++ news/src/Components/utils/formatDate.js
@@ -1,0 +1,6 @@
+export const formatDate = (date) => {
+    const splitByDayAndTime = date.split("T")
+    const splitDate = splitByDayAndTime[0].split("-")
+    const orderedDate = `${splitDate[2]}-${splitDate[1]}-${splitDate[0]}`
+    return orderedDate
+}
