+++ news/src/Components/utils/removeHtmlTags.js
@@ -1,0 +1,3 @@
+export const removeHtmlTags = (string) => {
+    return string.replace(/(<([^>]+)>)/ig, '* ')
+}
