+++ news/src/controllers/getScopeModifier.js
@@ -1,0 +1,6 @@
+export const getScopeModifier = (scope) =>{
+    if (scope === "everything")
+        { return "everything?q="
+    }
+    return "top-headlines?q="
+}
